feat: [#165] create views layer by renaming user_output to views
