+++ .github/copilot-instructions.md
@@ -99,7 +99,7 @@ These principles should guide all development decisions, code reviews, and featu
 
 9. **Before making engineering decisions**: Document significant architectural or design decisions as Architectural Decision Records (ADRs) in `docs/decisions/`. Read [`docs/decisions/README.md`](../docs/decisions/README.md) for the ADR template and guidelines. This ensures decisions are properly documented with context, rationale, and consequences for future reference.
 
-10. **When organizing code within modules**: Follow the module organization conventions in [`docs/contributing/module-organization.md`](../docs/contributing/module-organization.md). Use top-down organization with public items first, high-level abstractions before low-level details, and important responsibilities before secondary concerns like error types. **Prefer imports over full namespace paths** - always import types and use short names (e.g., `Arc<UserOutput>`) unless disambiguating name conflicts. Never use long paths like `std::sync::Arc<crate::presentation::user_output::UserOutput>` in regular code.
+10. **When organizing code within modules**: Follow the module organization conventions in [`docs/contributing/module-organization.md`](../docs/contributing/module-organization.md). Use top-down organization with public items first, high-level abstractions before low-level details, and important responsibilities before secondary concerns like error types. **Prefer imports over full namespace paths** - always import types and use short names (e.g., `Arc<UserOutput>`) unless disambiguating name conflicts. Never use long paths like `std::sync::Arc<crate::presentation::views::UserOutput>` in regular code.
 
 11. **When writing Markdown documentation**: Be aware of GitHub Flavored Markdown's automatic linking behavior. Read [`docs/contributing/github-markdown-pitfalls.md`](../docs/contributing/github-markdown-pitfalls.md) for critical patterns to avoid. **NEVER use hash-number patterns for enumeration or step numbering** - this creates unintended links to GitHub issues/PRs. Use ordered lists or alternative formats instead.
