+++ src/presentation/controllers/destroy/mod.rs
@@ -24,7 +24,7 @@
 //! use torrust_tracker_deployer_lib::bootstrap::Container;
 //! use torrust_tracker_deployer_lib::presentation::dispatch::ExecutionContext;
 //! use torrust_tracker_deployer_lib::presentation::controllers::destroy;
-//! use torrust_tracker_deployer_lib::presentation::user_output::VerbosityLevel;
+//! use torrust_tracker_deployer_lib::presentation::views::VerbosityLevel;
 //!
 //! let container = Container::new(VerbosityLevel::Normal);
 //! let context = ExecutionContext::new(Arc::new(container));
@@ -41,7 +41,7 @@
 //! use torrust_tracker_deployer_lib::bootstrap::Container;
 //! use torrust_tracker_deployer_lib::presentation::dispatch::ExecutionContext;
 //! use torrust_tracker_deployer_lib::presentation::controllers::destroy;
-//! use torrust_tracker_deployer_lib::presentation::user_output::VerbosityLevel;
+//! use torrust_tracker_deployer_lib::presentation::views::VerbosityLevel;
 //!
 //! let container = Container::new(VerbosityLevel::Normal);
 //! let context = ExecutionContext::new(Arc::new(container));
@@ -61,7 +61,7 @@
 //! use parking_lot::ReentrantMutex;
 //! use std::cell::RefCell;
 //! use torrust_tracker_deployer_lib::presentation::controllers::destroy;
-//! use torrust_tracker_deployer_lib::presentation::user_output::{UserOutput, VerbosityLevel};
+//! use torrust_tracker_deployer_lib::presentation::views::{UserOutput, VerbosityLevel};
 //! use torrust_tracker_deployer_lib::infrastructure::persistence::repository_factory::RepositoryFactory;
 //! use torrust_tracker_deployer_lib::shared::clock::SystemClock;
 //!
