+++ src/presentation/mod.rs
@@ -9,13 +9,13 @@
 //! The presentation layer follows a four-layer MVC architecture.
 //! This is part of [Presentation Layer Reorganization](../../docs/refactors/plans/presentation-layer-reorganization.md).
 //!
-//! **Progress**: 3/6 proposals completed (50%), Proposal #3 complete - ready for Proposal #4
+//! **Progress**: 4/6 proposals completed (67%), Proposal #4 complete - Views layer established
 //!
 //! ### Layer Architecture
 //!
 //! ```text
 //! Input → Dispatch → Controllers → Views
-//!   ✅       ✅         ✅         ⏳
+//!   ✅       ✅         ✅         ✅
 //! ```
 //!
 //! | Layer        | Status           | Purpose                                        |
@@ -23,7 +23,7 @@
 //! | **Input**    | ✅ Complete      | CLI argument parsing and validation           |
 //! | **Dispatch** | ✅ Complete      | Command routing and execution context         |
 //! | **Controllers** | ✅ Complete   | Command handling and business logic coordination |
-//! | **Views**    | ⏳ Planned       | Output formatting and presentation            |
+//! | **Views**    | ✅ Complete      | Output formatting and presentation            |
 //!
 //! ## Current Module Structure
 //!
@@ -65,8 +65,8 @@
 //! │   ├── tests/        # Controller layer integration tests
 //! │   └── mod.rs        # Controller layer exports
 //! │
-//! ├── user_output/      # ⏳ Future Views Layer (will be renamed to views/)
-//! │   └── ...           # Output formatting and presentation
+//! ├── views/           # ✅ Views Layer - Output formatting and presentation
+//! │   └── ...           # User interface output and formatting
 //! ├── progress.rs       # ⏳ Will move to views/progress/
 //! ├── errors.rs         # Unified error types for all commands
 //! └── mod.rs            # This file - layer exports and documentation
@@ -174,7 +174,7 @@ pub mod error;
 pub mod errors;
 pub mod input;
 pub mod progress;
-pub mod user_output;
+pub mod views;
 
 // Re-export commonly used presentation types for convenience
 pub use controllers::create::{
@@ -188,7 +188,7 @@ pub use error::handle_error;
 pub use errors::CommandError;
 pub use input::{Cli, Commands, GlobalArgs};
 pub use progress::ProgressReporter;
-pub use user_output::{Theme, UserOutput, VerbosityLevel};
+pub use views::{Theme, UserOutput, VerbosityLevel};
 
 #[cfg(test)]
 mod tests {
