+++ src/presentation/views/messages/result.rs
